backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
# psycopg[binary]==3.1.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.11.1
pytest==7.4.0
httpx==0.24.0
